# pkg: don't expand deps of restricting deps