@rgrinberg 2025-01-07
Is it possible to add a test to demonstrate a change in behavior? Since this change is loading opam files, I imagine that it should avoid raising an error when one of these opam files has an error.