@talex5 2025-01-09
Upstreaming would be good. It would likely make the solver service a fair bit faster.