@gridbugs 2025-01-22
> Upstreaming would be good. It would likely make the solver service a fair bit faster.

@talex5 I had a look at upstreaming this change but dune's vendored copy of 0install has diverged quite a bit from upstream. I don't have time right now to work out how to apply this change upstream.